% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ATLANTIC_OCEAN_RAW_GEOMETRY}
\alias{ATLANTIC_OCEAN_RAW_GEOMETRY}
\title{The core data for the ICCAT Atlantic Ocean area}
\format{
\describe{
  \item{CODE}{The area code}
  \item{TYPE_CODE}{The area type code}
  \item{NAME_EN}{The area English name}
  \item{NAME_ES}{The area Spanish name}
  \item{NAME_FR}{The area French name}
  \item{SURFACE_IN_ICCAT_AREA}{The extent of the intersection between the area and the ICCAT area}
  \item{GEOMETRY_WKT}{The geometry of the area}
}
}
\usage{
ATLANTIC_OCEAN_RAW_GEOMETRY
}
\description{
The core data for the ICCAT Atlantic Ocean area
}
\keyword{datasets}
