% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_constants.R
\name{geometries_for}
\alias{geometries_for}
\title{Converts into simple features a set of raw geometries as described (in _Well-Known-Text_ format, i.e., _WKT_) by their \code{GEOMETRY_WKT} column).
If required performs a coordinate reference system conversion before returning the simple features.}
\usage{
geometries_for(
  raw_geometries,
  source_crs = CRS_WGS84,
  target_crs = CRS_EQUIDISTANT
)
}
\arguments{
\item{raw_geometries}{a set of raw geometries with a \code{GEOMETRY_WKT} column storing the geometry definition in WKT format}

\item{source_crs}{a source CRS (defaults to \code{\link{CRS_WGS84}})}

\item{target_crs}{a target CRS (defaults to \code{\link{CRS_EQUIDISTANT}})}
}
\value{
a set of Simple Features built from the original raw geometries converted in the target CRS
}
\description{
Converts into simple features a set of raw geometries as described (in _Well-Known-Text_ format, i.e., _WKT_) by their \code{GEOMETRY_WKT} column).
If required performs a coordinate reference system conversion before returning the simple features.
}
