% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRIDS_5x5_RAW_GEOMETRIES}
\alias{GRIDS_5x5_RAW_GEOMETRIES}
\title{The core data for all 5x5 grids within the ICCAT area of competence}
\format{
\describe{
  \item{CODE}{The area code}
  \item{NAME_EN}{The area English name}
  \item{NAME_ES}{The area Spanish name}
  \item{NAME_FR}{The area French name}
  \item{ICCAT_AREA_INTERSECTION}{The extent of the intersection between the area and the ICCAT area}
  \item{GEOMETRY_WKT}{The geometry of the area}
}
}
\usage{
GRIDS_5x5_RAW_GEOMETRIES
}
\description{
The core data for all 5x5 grids within the ICCAT area of competence
}
\keyword{datasets}
