% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_base.R
\name{map.atlantic}
\alias{map.atlantic}
\title{Provides a base empty map for the Atlantic and Mediterranean areas}
\usage{
map.atlantic(
  xlim = DEFAULT_XLIM,
  x_breaks_every = 10,
  ylim = DEFAULT_YLIM,
  y_breaks_every = 10,
  axis_font_size = DEFAULT_AXIS_FONT_SIZE,
  fill_color = DEFAULT_FILL_COLOR,
  border_color = DEFAULT_BORDER_COLOR,
  border_line_width = DEFAULT_BORDER_LINE_WIDTH,
  force_show_all_atlantic = FALSE,
  crs = CRS_EQUIDISTANT,
  background_plot_function = NULL
)
}
\arguments{
\item{xlim}{a vector specifying the range (min - max) of longitudes to include in the map}

\item{x_breaks_every}{the distance between two consecutive breaks on the X axis}

\item{ylim}{a vector specifying the range (min - max) of latitudes to include in the map}

\item{y_breaks_every}{the distance between two consecutive breaks on the X axis}

\item{axis_font_size}{the axis font size}

\item{fill_color}{the color used to fill the land areas on the map}

\item{border_color}{the color used for the border of the land areas}

\item{border_line_width}{the line width for the border of the land areas}

\item{force_show_all_atlantic}{to ensure that the entire Atlantic ocean / ICCAT area is always displayed in full on the map}

\item{crs}{the target Coordinate Reference System used to plot the map. Defaults to \code{\link{CRS_EQUIDISTANT}}}

\item{background_plot_function}{an optional function used to initialize the background layer of the map}
}
\value{
a base empty map for the Atlantic and Mediterranean areas
}
\description{
Provides a base empty map for the Atlantic and Mediterranean areas
}
