% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_base.R
\name{map.coordinates_sf}
\alias{map.coordinates_sf}
\title{Creates a \code{\link{coord_sf}} object with the specified characteristics}
\usage{
map.coordinates_sf(
  xlim = NULL,
  ylim = NULL,
  source_crs = CRS_WGS84,
  target_crs = CRS_EQUIDISTANT
)
}
\arguments{
\item{xlim}{the vector with the two latitude limits (min - max)}

\item{ylim}{the vector with the two longitude limits (min - max)}

\item{source_crs}{the source Coordinate Reference System}

\item{target_crs}{the target Coordinate Reference System}
}
\value{
a \code{\link{coord_sf}} object that can be used to force the proper displaying of a map
}
\description{
Creates a \code{\link{coord_sf}} object with the specified characteristics
}
