% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_catdis.R
\name{map.heat.catdis}
\alias{map.heat.catdis}
\title{Produces a heatmap chart of CATDIS catch data by 5x5 grid}
\usage{
map.heat.catdis(
  base_map = map.atlantic(crs = iccat.pub.maps::CRS_EQUIDISTANT),
  catdis_data,
  gears_to_keep = NULL,
  gear,
  num_breaks = 5,
  crs = iccat.pub.maps::CRS_EQUIDISTANT
)
}
\arguments{
\item{base_map}{the base map to use}

\item{catdis_data}{the CATDIS data}

\item{gears_to_keep}{a vector of gears to keep in the final map. All gears with codes other than those provided will collapse in a generic _Other gears_ category}

\item{gear}{the gear whose CATDIS catch data should be used to produce the heatmap}

\item{num_breaks}{the number of breaks for the heatmap catch scale}

\item{crs}{the Coordinate Reference System to use}
}
\value{
a CATDIS heatmap showing catches of a given gear by magnitude and 5x5 grid
}
\description{
Produces a heatmap chart of CATDIS catch data by 5x5 grid
}
