% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_catdis.R
\name{map.pie.catdis.schooltype}
\alias{map.pie.catdis.schooltype}
\title{TBD}
\usage{
map.pie.catdis.schooltype(
  base_map = map.atlantic(crs = iccat.pub.maps::CRS_EQUIDISTANT),
  catdis_data,
  default_radius = pi,
  max_catch = NA,
  center_pies = TRUE,
  legend.x = -90,
  legend.y = -25,
  crs = iccat.pub.maps::CRS_EQUIDISTANT
)
}
\arguments{
\item{base_map}{TBD}

\item{catdis_data}{TBD}

\item{default_radius}{TBD}

\item{max_catch}{TBD}

\item{center_pies}{TBD}

\item{legend.x}{TBD}

\item{legend.y}{TBD}

\item{crs}{TBD}
}
\value{
TBD
}
\description{
TBD
}
