% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stocks.R
\name{map.stocks}
\alias{map.stocks}
\title{TBD}
\usage{
map.stocks(
  species_codes,
  stock_codes = NULL,
  base_map = map.atlantic(),
  fill_areas = TRUE,
  add_labels = TRUE
)
}
\arguments{
\item{species_codes}{TBD}

\item{stock_codes}{TBD}

\item{base_map}{TBD}

\item{fill_areas}{TBD}

\item{add_labels}{TBD}
}
\value{
TBD
}
\description{
TBD
}
