% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPECIES_TO_AREAS_MAPPINGS}
\alias{SPECIES_TO_AREAS_MAPPINGS}
\title{All areas (stock and sampling) associated to a given species by species code}
\format{
\describe{
  \item{SPECIES_CODE}{The species code}
  \item{STOCK_CODE}{The stock area code}
  \item{SAMPLING_AREA_CODE}{The sampling area code}
}
}
\usage{
SPECIES_TO_AREAS_MAPPINGS
}
\description{
All areas (stock and sampling) associated to a given species by species code
}
\keyword{datasets}
