% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STOCK_AND_SAMPLING_AREAS_RAW_GEOMETRIES}
\alias{STOCK_AND_SAMPLING_AREAS_RAW_GEOMETRIES}
\title{The core data for all stock and sampling areas}
\format{
\describe{
  \item{CODE}{The area code}
  \item{TYPE_CODE}{The area type code}
  \item{NAME_EN}{The area English name}
  \item{NAME_ES}{The area Spanish name}
  \item{NAME_FR}{The area French name}
  \item{SURFACE_IN_ICCAT_AREA}{The extent of the intersection between the area and the ICCAT area}
  \item{GEOMETRY_WKT}{The geometry of the area}
}
}
\usage{
STOCK_AND_SAMPLING_AREAS_RAW_GEOMETRIES
}
\description{
The core data for all stock and sampling areas
}
\keyword{datasets}
