% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_constants.R
\name{geometries_for}
\alias{geometries_for}
\title{TBD}
\usage{
geometries_for(
  raw_geometries,
  source_crs = CRS_WGS84,
  target_crs = CRS_EQUIDISTANT
)
}
\arguments{
\item{raw_geometries}{TBD}

\item{source_crs}{TBD}

\item{target_crs}{TBD}
}
\value{
TBD
}
\description{
TBD
}
