% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRIDS_5x5_RAW_GEOMETRIES}
\alias{GRIDS_5x5_RAW_GEOMETRIES}
\title{The core data for all 5x5 grids within the ICCAT area of competence}
\format{
\describe{
  \item{CODE}{The area code}
  \item{TYPE_CODE}{The area type code}
  \item{CENTER_LAT}{The area center latitude}
  \item{CENTER_LON}{The area center longitude}
  \item{GEOMETRY_WKT}{The geometry of the area}
}
}
\usage{
GRIDS_5x5_RAW_GEOMETRIES
}
\description{
The core data for all 5x5 grids within the ICCAT area of competence
}
\keyword{datasets}
