% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_base.R
\name{map.atlantic}
\alias{map.atlantic}
\title{TBD}
\usage{
map.atlantic(
  xlim = DEFAULT_XLIM,
  x_breaks_every = 10,
  ylim = DEFAULT_YLIM,
  y_breaks_every = 10,
  axis_font_size = DEFAULT_AXIS_FONT_SIZE,
  fill_color = DEFAULT_FILL_COLOR,
  border_color = DEFAULT_BORDER_COLOR,
  border_line_width = DEFAULT_BORDER_LINE_WIDTH
)
}
\arguments{
\item{xlim}{TBD}

\item{x_breaks_every}{TBD}

\item{ylim}{TBD}

\item{y_breaks_every}{TBD}

\item{axis_font_size}{TBD}

\item{fill_color}{TBD}

\item{border_color}{TBD}

\item{border_line_width}{TBD}
}
\value{
TBD
}
\description{
TBD
}
